requirements.txt
langchain
langchain-groq
crewai
python-dotenv
